unit tests

+++ app/Actions/Discount/DeleteDiscount.php
@@ -2,8 +2,8 @@
 
 namespace App\Actions\Discount;
 
-use App\Models\Campaign;
 use App\Models\Discount;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Lorisleiva\Actions\Concerns\AsAction;
 
 class DeleteDiscount
@@ -12,6 +12,9 @@ class DeleteDiscount
 
     public function handle(Discount $discount): Discount
     {
+        if (!is_null($discount->deleted_at)) {
+            throw new ModelNotFoundException();
+        }
         $discount->delete();
         $discount->products()->detach($discount->products()->pluck('id')->toArray());
         return $discount->refresh();
