+++ app/Actions/Product/DeleteProduct.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Product;
 
 use App\Models\Product;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Lorisleiva\Actions\Concerns\AsAction;
 
 class DeleteProduct
@@ -11,6 +12,9 @@ class DeleteProduct
 
     public function handle(Product $product): Product
     {
+        if (!is_null($product->deleted_at)) {
+            throw new ModelNotFoundException();
+        }
         $product->delete();
         return $product->refresh();
     }
